requirements.txt
pytest~=8.4.1
playwright~=1.54.0
allure-pytest
requests~=2.32.4
pydantic~=2.11.7
python-dotenv~=1.1.1
allure-python-commons~=2.15.0